Add authentication
In this CL was added functionality that allows to sign up and
sign in users. To sign up client should send POST request to host/api/SignUp
server gives response like status that notify that user was added or not.
Also to sign in client should send POST request to host/api/Login in case of
success server gives response with token of user otherwise empty token.